Refactored index.ts
refactored index.ts to individual commands.
