+++ src/commands/templates/templateTypes.ts
@@ -1,0 +1,7 @@
+import * as vscode from 'vscode';
+
+export interface TemplateTypeOption extends vscode.QuickPickItem {
+    templateKey: string;
+    defaultFilename: string;
+    displayName: string;
+}
